Update the Testcases + Travis Integration
Tests now have a seperate workspace for easier execution and to make them more robust. Currently the tests are running in a third Travis Build, next to the maven and yarn builds. However, since the tests require maven and yarn to be run, these tests could be put together in the future.

Right now two tests are failing due to #103.

Signed-off-by: Simon Graband <simon.graband@tum.de>